Remove util dir
