+++ apps/backend/src/api_clients/paper_meta.rs
@@ -1,5 +1,4 @@
-use crate::config::PAPER_API_URL;
-use crate::util::request::get_and_format;
+use crate::{api_clients::get_and_format, config::PAPER_API_URL};
 use serde::Deserialize;
 
 // API Types
