+++ apps/backend/src/models/files/binaries_lockfile.rs
@@ -2,7 +2,7 @@ use std::{collections::HashMap, path::PathBuf};
 
 use serde::{Deserialize, Serialize};
 
-use crate::{models::game::Game, util::hash::HashAlgorithm};
+use crate::{models::game::Game, models::hash::HashAlgorithm};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct BinaryLockfile {
