+++ apps/backend/src/models/hash.rs
