+++ apps/backend/src/models/mod.rs
@@ -1,4 +1,5 @@
 pub mod files;
 pub mod game;
+pub mod hash;
 pub mod secrets;
 pub mod server;
