--- apps/backend/src/util/mod.rs
@@ -1,3 +1,0 @@
-pub mod download;
-pub mod hash;
-pub mod request;
